Replaced int64_t with size_t

+++ src/fdl.cpp
@@ -1,6 +1,5 @@
 #include "fdl.h"
 #include <cmath>
-#include <cstdint>
 
 
 
@@ -8,7 +7,7 @@ const float MAX_VELOCITY_MAGNITUDE_SQR = 10.0f;
 
 
 
-#define ELEM(data, cols, i, j) (data)[(int64_t)(i) * (cols) + (j)]
+#define ELEM(data, cols, i, j) (data)[(size_t)(i) * (cols) + (j)]
